backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
google-genai>=1.0.0
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.2.1
google-api-python-client==2.116.0
google-auth==2.27.0
gspread==5.12.0
oauth2client==4.1.3
pytest==8.0.0
pytest-mock==3.12.0
httpx==0.26.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.0
rich==13.7.0
python-json-logger==2.0.7
sqlalchemy==2.0.25
